home section refactor
